Remove Peoplefinder namespace from models
