+++ app/models/concerns/notifications.rb
@@ -1,12 +1,10 @@
-require 'peoplefinder'
-
-module Peoplefinder::Concerns::Notifications
+module Concerns::Notifications
   extend ActiveSupport::Concern
 
   included do
     def send_create_email!(current_user)
       if should_send_email_notification?(email, current_user)
-        Peoplefinder::UserUpdateMailer.new_profile_email(
+        UserUpdateMailer.new_profile_email(
           self, current_user.email
         ).deliver_later
       end
@@ -14,7 +12,7 @@ module Peoplefinder::Concerns::Notifications
 
     def send_update_email!(current_user)
       if should_send_email_notification?(email, current_user)
-        Peoplefinder::UserUpdateMailer.updated_profile_email(
+        UserUpdateMailer.updated_profile_email(
           self, current_user.email
         ).deliver_later
       end
@@ -22,7 +20,7 @@ module Peoplefinder::Concerns::Notifications
 
     def send_destroy_email!(current_user)
       if should_send_email_notification?(email, current_user)
-        Peoplefinder::UserUpdateMailer.deleted_profile_email(
+        UserUpdateMailer.deleted_profile_email(
           self, current_user.email
         ).deliver_later
       end
@@ -31,7 +29,7 @@ module Peoplefinder::Concerns::Notifications
   private
 
     def should_send_email_notification?(email, current_user)
-      Peoplefinder::EmailAddress.new(email).valid_address? &&
+      EmailAddress.new(email).valid_address? &&
         current_user.email != email
     end
   end
