--- app/models/peoplefinder/permitted_domain.rb
@@ -1,6 +1,0 @@
-module Peoplefinder
-  class PermittedDomain < ActiveRecord::Base
-    self.table_name = 'permitted_domains'
-    validates :domain, presence: true
-  end
-end
