+++ app/models/permitted_domain.rb
@@ -1,0 +1,3 @@
+class PermittedDomain < ActiveRecord::Base
+  validates :domain, presence: true
+end
